AST

Description

Module for Abstract Syntax Trees (ASTs) that are Trees with a fixed node type.

Documentation

data Node g a

Node type of ASTs.

Constructors

Node

Fields

value :: a

The value that is encapsulated by a node. Most of the time this is a String referring to source code statements.

grammartype :: g

The grammar rule this node was parsed from. g should be an instance of Grammar.

uuid :: UUID

A unique identifier to trace nodes across several versions of ASTs.

type AST g a = Tree (Node g a)

Type for abstract syntax trees (ASTs) representing source code. It is a tree in which each node represents a source code entity by a value, a grammar type and a UUID.

optionaltype :: Grammar g => Node g a -> NodeType

Returns the node type of a nodes grammar type.

node :: Grammar g => a -> g -> State UUID (Node g a)

Creates a new node from a value and a type by generating a new UUID for it.

uuidOf :: AST g a -> UUID

Returns the UUID of an ASTs root.

findById :: UUID -> AST g a -> Maybe (AST g a)

Finds a subtree in the given AST whose root has the given UUID. Returns Nothing iff no such subtree exists.

findByValue :: Eq a => a -> AST g a -> Maybe (AST g a)

Finds a subtree in the given AST whose root has the given value. Returns Nothing iff no such subtree exists.

findByGrammarType :: Eq g => g -> AST g a -> Maybe (AST g a)

Finds a subtree in the given AST whose root has the given grammar type. Returns Nothing iff no such subtree exists.

optionalAncestors :: (Eq a, Grammar g) => AST g a -> AST g a -> [AST g a]

Returns all ancestors of a given subtree (second argument) in the given tree (first argument), that are optional (i.e., their optionaltype is Optional).