ASTPrettyPrinter

Description

Module for pretty printing of ASTs.

Documentation

class Show g => ASTPrettyPrinter g where

An ASTPrettyPrinter can pretty print ASTs of a certain grammar.

Methods

prettyPrint :: Monoid b => b -> (Node g a -> Int -> b) -> (Node g a -> String -> b) -> (Node g a -> b) -> AST g a -> b

Pretty prints the given AST (last argument) to another type b (e.g., String). Parameters are: * the start indent (e.g., ""), * an indent generator that can generate an indent of a certain width within the context of printing a certain node (e.g., (_ i -> genIndent i)), * a function to lift strings to the output type b within the context of printing a certain node, * a function print nodes, and * the AST to print.

showCode :: (Show a, Grammar g, ASTPrettyPrinter g) => AST g a -> String

Default implementation for pretty printing an AST to a String.

showCodeAs :: (Monoid b, ASTPrettyPrinter g) => b -> (Node g a -> Int -> b) -> (Node g a -> String -> b) -> (Node g a -> b) -> AST g a -> b

Alias for prettyPrint.