Grammar

Description

Module for representing context-free grammars.

Documentation

data NodeType

Classification of nodes for feature traces and presence condition. (See Section 3.1 in the paper.)

Constructors

Mandatory

Optional

Wrapper

class Show g => Grammar g where

Type class for context-free grammars.

Methods

nodetypeof :: g -> NodeType

Tells for each grammar symbol which node type it has (e.g., if it is Mandatory in the AST).