ListUtil

Description

A collection of functions that modify lists.

Documentation

safehead :: [a] -> Maybe a

Returns the head of the given list as Just or Nothing, iff the list is empty.

getRange :: Int -> Int -> [a] -> [a]

Returns a continuous sublist within the given range [i, j].

insertAtIndex :: Int -> a -> [a] -> [a]

Inserts the given element into the given list at the given index. Subsequent elements will be pushed by one.

insertListAtIndex :: Int -> [a] -> [a] -> [a]

Inserts all elements from the first list into the second list at the given index. Similar to insertAtIndex but inserts all elements of a list rather than just a single element.

removeIndex :: Int -> [a] -> [a]

Removes the element at given index from the given list.

removeRange :: Int -> Int -> [a] -> [a]

Removes all elements within the given range [i, j] from the given list.

removeDuplicates :: Eq a => [a] -> [a]

Removes all duplicates from the given list. If two elements are equal (w.r.t. to Eq), the first one will be kept and the second one discarded.