Logic

Description

Type class for reasoning on Logics.

Documentation

class Logic l where

Type class to reason on logics. We use this type class to reason on propositional logic and the ternary logic by Sobocinski we use for feature trace recording.

Minimal complete definition

ltrue, lfalse, lnot, land, leval

Methods

ltrue :: l

The atomic value representing true in this logic.

lfalse :: l

The atomic value representing false in this logic.

lvalues :: [l]

A list of all atomic values of this logic. Default implementation comprises ltrue and lfalse.

lnot :: l -> l

Negation of a logical formula.

land :: [l] -> l

Conjunction of a list of logical formulas.

lor :: [l] -> l

Disjunction of a list of logical formulas.

limplies :: l -> l -> l

Implication between two logical formulas. The first argument p is one the left side of the implication and the second argument q is on the right (i.e., p => q).

lequals :: l -> l -> l

Equivalence between two logical formulas.

leval :: (l -> l) -> l -> l

Evaluates a logical formula. Arguments are

a function assigning variables to values

a formula to evaluate This function should return an element of lvalues.