NullPropositions

Description

Definition and operations on the ternary logic with null. In the paper, we call formulas of this logic nullable propositional formulas. Reuses PropositionalFormulas.

Documentation

type NullableFormula a = Maybe (PropositionalFormula a)

Data type for the ternary logic by Sobocinski. The Nothing case represents null as used in our paper.

isnull :: NullableFormula a -> Bool

Returns true iff the given formula is the value null.

notnull :: NullableFormula a -> Bool

Returns false iff the given formula is the value null.

assure :: NullableFormula a -> PropositionalFormula a

Converts the given nullable formula to a propositional formula, assuming that the given formula is not null. Crashes otherwise.

nullable_simplify :: NullableFormula a -> NullableFormula a

Simplifies the given formula. Uses simplify.

prettyPrint :: Show a => Maybe a -> String

Pretty Printing for nullable objects such as the nullable propositional logic.

Orphan instances

Logic a => Logic (Maybe a)
Any Logic can be lifted to a logic on Maybe. This adds a new value Nothing to the values of the given logic. In particular, NullableFormula is thus a Logic.

Instance details

Methods

ltrue :: Maybe a

lfalse :: Maybe a

lvalues :: [Maybe a]

lnot :: Maybe a -> Maybe a

land :: [Maybe a] -> Maybe a

lor :: [Maybe a] -> Maybe a

limplies :: Maybe a -> Maybe a -> Maybe a

lequals :: Maybe a -> Maybe a -> Maybe a

leval :: (Maybe a -> Maybe a) -> Maybe a -> Maybe a