StackPopBob

Description

Module for reproducing Bob's part of our motivating example. Bob propagates applicable edits by Alice to his variant. The example is described in detail in Section 2.2 of the paper and shown in Figure 3.

Documentation

featureColourPalette :: MonadColorPrinter m => FeatureFormulaColourPalette m -> FeatureFormulaColourPalette m

Colours for features and feature formulas used in this example. We chose terminal colours as close the the colours used in the paper as possible.

example :: MonadColorPrinter m => State UUID (Example m SimpleJavaGrammar String)

Example replaying our Bob's part of our motivating example shown in Figure 3 and described in Section 2.2 in our paper. Bob propagates Alice's changes on the pop method to his variant. This example directly reuses the first two edits of Alice and appends an artifical noop edit (edit_trace_only) to update the feature mappings that were recorded upon Alice's edits that were not applicable to Bob's variant but induced further feature mappings.