Tree

Description

Implementation of a rose tree (i.e., a tree whose nodes can have an arbitrary amount of children, including 0).

Documentation

data Tree a

Rose tree of elements of type a where children of nodes are given as lists.

Constructors

Tree a [Tree a]

isleaf :: Tree a -> Bool

Returns true iff the given tree is a leaf (i.e., it has no children).

element :: Tree a -> a

Returns the value stored in the root of the given tree.

tree :: (Eq a, Show a) => Tree a -> a -> Tree a

Finds a subtree in the given tree (first argument) whose root has the given value (second argument). Throws an error iff no such tree exists.

safetree :: Eq a => Tree a -> a -> Maybe (Tree a)

Same as tree but returns Nothing in the error case (i.e., when no subtree contains the given value).

toset :: Ord a => Tree a -> Set a

Transforms a tree into a set. The returned set contains exactly the values previously held in the tree.

find :: (Tree a -> Bool) -> Tree a -> Maybe (Tree a)

Find a subtree in the given tree (first argument) whose root matches the predicate (second argument).

findByNode :: (a -> Bool) -> Tree a -> Maybe (Tree a)

Same as find but takes a predicate over elements instead of a predicate over trees to identify the subtree of interest.

parent :: Eq a => Tree a -> Tree a -> Maybe (Tree a)

Returns the parent of the given node (second argument) in the given tree (first argument). The returned parent is a subtree of the first given tree and has the second given tree as child. Throws an error iff no parent exists.

ancestors :: Eq a => Tree a -> Tree a -> [Tree a]

manipulate :: (Tree a -> Tree a) -> Tree a -> Tree a

Applies the given function to each node from bottom to top (i.e., leaves are converted first, root last).

filterTrees :: (Tree a -> Bool) -> Tree a -> Tree a

filterNodes :: (Tree a -> Bool) -> Tree a -> Tree a

prettyPrint :: (Show a, Monoid b) => Int -> (String -> b) -> (a -> b) -> Tree a -> b

Pretty print function that folds the tree into a monoidial type b. b is the type that is the output of the print (e.g., String or Text). First argument is the length of the indent in spaces (e.g., 2 will produce an indent of " " per level). Second argument is a function that lifts strings to the output type b (e.g., a constructor of b accepting a string). Third argument is a function to convert tree elements to the output type (e.g., show if b is String).