Util

Description

A module containing various utility functions.

Documentation

genIndent :: Int -> String

Generates a string of i spaces where i is the given indent.

safeFromJust :: Maybe a -> (() -> a) -> a

Returns x iff the given Maybe is Just x. Otherwise, returns the value produced by the given generator function.

reversefoldr :: (a -> b -> b) -> b -> [a] -> b

Folds the given list after reversing it.

parenIf :: Bool -> String -> String

Generates parenthesis around the given string iff the given bool is true.

takeIf :: (a -> Bool) -> Maybe a -> Maybe a

Filters the given Maybe. If the maybe's element satisfies the given predicate, the element will be kept. Otherwise, returns Nothing.

nothingIf :: (a -> Bool) -> a -> Maybe a

Lifts a value to a Maybe based on a predicate. Iff the element satisfies the predicate, the result is Nothing. Otherwise returns the Just the element.

removeFirstAndLast :: [a] -> [a]

Removes the first and the last element of a list.

removeQuotes :: String -> String

If the given string starts and ends with quotes " (i.e., it is of the form ""something""), those parenthesis will be removed (i.e., turned to just "something"). This is used for showing strings from a polymorphic context.