Main

Documentation

data CodePrintStyle

Constructors

ShowAST

ShowCode

ShowTikz

data TraceDisplay

Constructors

Trace

PC

data TraceStyle

Constructors

Text

Colour

None

data OutputFormat

Constructors

OutputFormat

Fields

codeStyle :: CodePrintStyle

traceDisplay :: TraceDisplay

traceStyle :: TraceStyle

withTraceLines :: Bool

userFormat :: OutputFormat

The perspective of the developer who is editing code while traces are recorded in the background This is the format used in the figures in the paper.

userFormatDetailed :: OutputFormat

A variation of userFormat where traces and presence conditions can be investigated seperately at the same time. Code is coloured in the colour of its trace while presence conditions are indicated by coloured lines on the left.

astFormat :: OutputFormat

Shows the Abstract Syntax Tree of the source code with feature traces as formulas.

tikzFormat :: OutputFormat

Tikz export of AST with traces. Used for figures in the paper.

main :: IO ()

showExamples :: IO ()

headline :: MonadColorPrinter m => String -> m ()

finalizeExample :: State UUID (Example m g a) -> Example m g a

runStepwise :: (MonadColorPrinter m, Grammar g, ASTPrettyPrinter g) => OutputFormat -> State UUID (Example m g String) -> m ()

printASTWithTrace :: (MonadColorPrinter m, Grammar g, ASTPrettyPrinter g, Show a, Eq a) => OutputFormat -> FeatureFormulaColourPalette m -> AST g a -> FeatureTrace g a -> Doc (Attribute m)

printTraces :: (MonadColorPrinter m, Grammar g, ASTPrettyPrinter g, Show a, Eq a) => OutputFormat -> Example m g a -> [(FeatureTrace g a, AST g a)] -> Doc (Attribute m)

alsoShowInitialStateInHistory :: History g a -> History g a

propositional_values :: [PropositionalFormula String]

nullableFormula_values :: [NullableFormula String]

showTruthtables :: IO ()